fix(prettier-plugin-jsdoc): remove ifElse
Using formatStringLiterals with a R.ifElse meant that the getReducer function was being executed
every time, no matter if there were no string literals on the type.
The alternative could've been using R.curry with getReducer, but that would make the parsing of
the options and the composition for the formatter to happen on each iteration.

I prefer using a non Ramda function and parsing the options and creating the formatter only
once per type.